feat: move selectParamFromFilter into state/selectors
