+++ src/state/slices/portfolioSlice/selectors.ts
@@ -32,6 +32,8 @@ import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import type { ReduxState } from 'state/reducer'
 import { createDeepEqualOutputSelector } from 'state/selector-utils'
+import { selectParamFromFilterOptional } from 'state/selectors'
+import { selectParamFromFilter } from 'state/selectors'
 import { selectAssets } from 'state/slices/assetsSlice/selectors'
 import {
   selectFarmContractsFiatBalance,
@@ -65,38 +67,6 @@ import type {
   StakingDataByValidatorId,
 } from './portfolioSliceCommon'
 import { findAccountsByAssetId } from './utils'
-
-type ParamFilter = {
-  assetId: AssetId
-  accountId: AccountId
-  accountNumber: number
-  chainId: ChainId
-  validatorAddress: PubKey
-}
-type OptionalParamFilter = {
-  assetId?: AssetId
-  accountId?: AccountId
-  validatorAddress?: PubKey
-}
-type ParamFilterKey = keyof ParamFilter
-type OptionalParamFilterKey = keyof OptionalParamFilter
-
-const selectParamFromFilter = <T extends ParamFilterKey>(param: T) =>
-  createCachedSelector(
-    (_state: ReduxState, filter: Pick<ParamFilter, T>): ParamFilter[T] | '' =>
-      filter?.[param] ?? '',
-    param => param,
-  )((_state: ReduxState, filter: Pick<ParamFilter, T>) => filter?.[param] ?? param)
-const selectParamFromFilterOptional = <T extends OptionalParamFilterKey>(param: T) =>
-  createCachedSelector(
-    (_state: ReduxState, filter: Pick<OptionalParamFilter, T>): OptionalParamFilter[T] | '' =>
-      filter?.[param] ?? '',
-    param => param,
-  )(
-    (_state: ReduxState, filter: Pick<OptionalParamFilter, T>) =>
-      `${param}-${filter?.[param]}` ?? param,
-  )
-
 // We should prob change this once we add more chains
 const FEE_ASSET_IDS = [
   ethAssetId,
